requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
nhanes-dl>=0.1.0
seaborn>=0.12.0
matplotlib>=3.6.0
umap-learn>=0.5.0



